scripts/requirements.txt
requests==2.31.0
SPARQLWrapper==2.0.0
openai==1.54.0
Pillow==10.4.0
python-slugify==8.0.1
pandas==2.1.4
numpy==1.24.3
pillow-heif==0.19.0
python-dotenv==1.0.0
tqdm==4.66.1
beautifulsoup4==4.12.2
